Factor out follow test (#268)
* Factor out follow test

* Refactor to test function
