+++ tests/index.test.js
@@ -1,4 +1,5 @@
 const createSdk = require('./utils/create-sdk')
+const testFollowMethod = require('./test-follow-method')
 
 const {
   existsSync,
@@ -24,6 +25,8 @@ const localDHT = async () => {
   dht = node
   dhtBootstrap = url
 }
+
+testFollowMethod()
 
 test('ready', async t => {
   const p2p = createSdk()
@@ -233,39 +236,6 @@ test('saveItem: should throw ValidationError with invalid metadata', async t => 
   } catch (err) {
     t.fail(err)
   }
-  await p2p.destroy()
-  t.end()
-})
-
-test('follows: must not self-reference', async t => {
-  const p2p = createSdk()
-
-  t.plan(1)
-
-  const sampleProfile = {
-    type: 'profile',
-    title: 'professorX',
-    subtype: '',
-    avatar: './test.png',
-    follows: [
-      'f7daadc2d624df738abbccc9955714d94cef656406f2a850bfc499c2080627d4'
-    ],
-    contents: [
-      '00a4f2f18bb6cb4e9ba7c2c047c8560d34047457500e415d535de0526c6b4f23+12'
-    ]
-  }
-
-  const { rawJSON: profile } = await p2p.init(sampleProfile)
-
-  try {
-    await p2p.follow(profile.url, profile.url)
-  } catch (err) {
-    t.ok(
-      err instanceof SDK.errors.ValidationError,
-      'should throw when tries to self-reference'
-    )
-  }
-
   await p2p.destroy()
   t.end()
 })
